# Refactoring & getTreeInfo speedup