--- a/src/NITTA/Synthesis/Analysis.hs
+++ b/src/NITTA/Synthesis/Analysis.hs
@@ -17,7 +17,6 @@ import Control.Concurrent.STM
 import Data.HashMap.Strict qualified as HM
 import GHC.Generics
 import NITTA.Model.TargetSystem (processDuration)
-import NITTA.Synthesis.Explore (isComplete, isLeaf)
 import NITTA.Synthesis.Types
 
 -- | Metrics of synthesis tree process