--- a/src/NITTA/Synthesis/Analysis.hs
+++ b/src/NITTA/Synthesis/Analysis.hs
@@ -1,5 +1,3 @@
-{-# LANGUAGE BangPatterns #-}
-
 {- |
 Module      : NITTA.Synthesis.Analysis
 Description : Analysis synthesis proccess.