--- a/src/NITTA/Synthesis/Explore.hs
+++ b/src/NITTA/Synthesis/Explore.hs
@@ -44,12 +44,15 @@ synthesisTreeRootIO = atomically . rootSynthesisTreeSTM
 
 rootSynthesisTreeSTM model = do
     sSubForestVar <- newEmptyTMVar
+    let sState = nodeCtx Nothing model
     return
         Tree
             { sID = def
-            , sState = nodeCtx Nothing model
+            , sState
             , sDecision = Root
             , sSubForestVar
+            , isLeaf = isLeaf' sState
+            , isComplete = isComplete' sState
             }
 
 -- | Get specific by @nId@ node from a synthesis tree.
@@ -105,22 +108,20 @@ subForest (objective function value less than zero).
 -}
 positiveSubForestIO tree = filter ((> 0) . defScore . sDecision) <$> subForestIO tree
 
-isLeaf
-    Tree
-        { sState =
-            SynthesisState
-                { sAllocationOptions = []
-                , sBindOptions = []
-                , sDataflowOptions = []
-                , sBreakLoopOptions = []
-                , sResolveDeadlockOptions = []
-                , sOptimizeAccumOptions = []
-                , sConstantFoldingOptions = []
-                }
+isLeaf'
+    SynthesisState
+        { sAllocationOptions = []
+        , sBindOptions = []
+        , sDataflowOptions = []
+        , sBreakLoopOptions = []
+        , sResolveDeadlockOptions = []
+        , sOptimizeAccumOptions = []
+        , sConstantFoldingOptions = []
         } = True
-isLeaf _ = False
+isLeaf' _ = False
 
-isComplete = isSynthesisComplete . sTarget . sState
+isComplete' = isSynthesisComplete . sTarget
+
 -- * Internal
 
 exploreSubForestVar parent@Tree{sID, sState} =
@@ -142,6 +143,8 @@ exploreSubForestVar parent@Tree{sID, sState} =
                     , sState = ctx'
                     , sDecision = desc
                     , sSubForestVar
+                    , isLeaf = isLeaf' ctx'
+                    , isComplete = isComplete' ctx'
                     }
 
 decisionAndContext parent@Tree{sState = ctx} o =