--- a/src/NITTA/Synthesis/Explore.hs
+++ b/src/NITTA/Synthesis/Explore.hs
@@ -17,8 +17,6 @@ module NITTA.Synthesis.Explore (
     getTreePathIO,
     subForestIO,
     positiveSubForestIO,
-    isComplete,
-    isLeaf,
 ) where
 
 import Control.Concurrent.STM