`stack build && stack exec nitta -- -p=8080 -t=fx32.32 examples/sin_ident/sin_ident.lua`

Performance of `time http GET localhost:8080/treeInfo`:

1. Before optimisation:
    
    ```shell
    Executed in   19.20 secs      fish           external
    ```

2. After predicate rewriting (https://github.com/ryukzak/nitta/commit/f262ad3e5dbfd3f665eb543145f57763b734e2a8):

    ```shell
    Executed in    2.02 secs      fish           external
    ```

3. After Force strictness (https://github.com/ryukzak/nitta/commit/c80ae1ad83d951907f335b126d4c2fb9d57b54ae):

    ```shell
    Executed in    1.93 secs      fish           external
    ```

4. After extraction isLeaf and isComplete to tree data (https://github.com/ryukzak/nitta/commit/8020ccbb194958c218ad35c7db130419eaa44fe9):

    ```shell
    Executed in    1.86 secs      fish           external
    ```

close #153