@iburakov 2023-06-09
This was the problem `isComplete` and `isLeaf` in ML branch

<img src="https://github.com/ryukzak/nitta/assets/5229130/bf89ef53-6d87-4c22-862e-c9bf88ec9659" width="50%">

If I get it right, in the `ViewHelper` module we can now take raw data from `Tree` with `isLeaf` and `isComplete` from `NITTA.Synthesis.Types`, so it shouldn't be a problem, no this kind of cyclic dependency is reintroduced.